+++ efficax-server/efficax-server/src/network/packet.rs
@@ -6,12 +6,12 @@ use super::data::NetworkData;
 pub struct NetworkPacket {
     pub is_tcp: bool,
     pub addrs: Vec<SocketAddr>,
-    pub tick_id: u8,
+    pub tick_id: u32,
     pub data: NetworkData
 }
 
 impl NetworkPacket {
-    pub fn unicast(is_tcp: bool, addr: SocketAddr, tick_id: u8, data: NetworkData) -> Self {
+    pub fn unicast(is_tcp: bool, addr: SocketAddr, tick_id: u32, data: NetworkData) -> Self {
         NetworkPacket {
             is_tcp,
             addrs: vec![addr],
@@ -20,7 +20,7 @@ impl NetworkPacket {
         }
     }
 
-    pub fn multicast(is_tcp: bool, addrs: Vec<SocketAddr>, tick_id: u8, data: NetworkData) -> Self {
+    pub fn multicast(is_tcp: bool, addrs: Vec<SocketAddr>, tick_id: u32, data: NetworkData) -> Self {
         NetworkPacket {
             is_tcp,
             addrs,
