+++ efficax-server/efficax-utils/src/id_gen.rs
@@ -1,15 +1,15 @@
 pub struct IdGen {
-    next_id: u64
+    next_id: u32
 }
 
 impl IdGen {
-    pub fn new(first_id: u64) -> Self {
+    pub fn new(first_id: u32) -> Self {
         IdGen {
             next_id: first_id
         }
     }
 
-    pub fn get(&mut self) -> u64 {
+    pub fn get(&mut self) -> u32 {
         let id = self.next_id;
         self.next_id += 1;
         id
